+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### TestAdapter
 - Update `TestDiscoverer` with `Log` feature
 - Update `RevitTestConsole` with `RunTestReadWithLog` feature
-- Add `ricaun.Security.WinTrust` package
+- Add `ricaun.Security.WinTrust` package with `PrivateAssets` 
 - Add `TrustApplicationUtils` to check if application is trusted
 - Add `DelayAfterExit` in `ProcessStart` with value `100`
 ### Fixed